[NHDR-308] refactor: 함수 파일 따로 관리
