+++ src/api/nohoo/gold.ts
@@ -4,7 +4,7 @@ import type { GoldPrice } from '@/types/nohoo/gold';
 const API_KEY = import.meta.env.VITE_ALPHA_VANTAGE_KEY;
 const BASE_URL = 'https://www.alphavantage.co/query';
 
-export const fetchGoldPrice = async (): Promise<GoldPrice[]> => {
+export async function fetchGoldPrice(): Promise<GoldPrice[]> {
   const url = `${BASE_URL}?function=TIME_SERIES_DAILY&symbol=XAUUSD&apikey=${API_KEY}`;
   const res = await axios.get(url);
   const rawData = res.data['Time Series (Daily)'];
@@ -19,4 +19,4 @@ export const fetchGoldPrice = async (): Promise<GoldPrice[]> => {
     .sort((a, b) => new Date(a.date).getTime() - new Date(b.date).getTime()); // 날짜 오름차순 정렬
 
   return prices;
-};
+}
