+++ src/api/products/productDetail.ts
@@ -4,14 +4,16 @@ import type {
   FundDailyReturn,
 } from '@/types/products/productDetail';
 
+// 금융 상품 상세 조회
 export async function fetchProductDetail(
   id: string
 ): Promise<ProductDetailApiResponse> {
   const res = await api.get(`/api/retirement/${id}`);
-  console.log('Product detail fetched:', res.data);
+  // console.log('Product detail fetched:', res.data);
   return res.data;
 }
 
+// 펀드 일일 수익률 조회
 export async function fetchFundReturn(
   id: string,
   range = '3m'
@@ -20,6 +22,6 @@ export async function fetchFundReturn(
     `/api/retirement/${id}/returns`,
     { params: { range } }
   );
-  console.log('Fund return fetched:', res.data);
+  // console.log('Fund return fetched:', res.data);
   return res.data ?? [];
 }
