+++ src/api/user/preferences.ts
@@ -1,13 +1,5 @@
 import api from '@/api';
-
-// 투자성향 질문 답변 데이터 타입
-export interface PreferencesRequest {
-  q1: string;
-  q2: string;
-  q3: string;
-  q4: string;
-  q5: string;
-}
+import type { PreferencesRequest } from '@/types/user/preferences';
 
 // 투자성향 질문 관련 API 함수들
 export const preferencesApi = {
