+++ src/types/event/event.ts
@@ -1,0 +1,12 @@
+interface Event {
+  news: {
+    category: number;
+    title: string;
+    summary: string;
+    link: string;
+    date: string;
+  };
+  point: number;
+}
+
+export type { Event };
