+++ src/types/user/preferences.ts
@@ -1,0 +1,10 @@
+// 투자성향 질문 답변 데이터 타입
+interface PreferencesRequest {
+  q1: string;
+  q2: string;
+  q3: string;
+  q4: string;
+  q5: string;
+}
+
+export type { PreferencesRequest };
